backend/requirements.txt
Django==5.1.7
django-background-tasks==1.2.8
django-cors-headers==4.4.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.9
requests==2.31.0